+++ mobile-app/app/_layout.tsx
@@ -6,7 +6,6 @@ import { useEffect } from 'react';
 import { useColorScheme } from 'react-native';
 import Onboarding from './Onboarding';
 import 'react-native-gesture-handler';
-import Login from './login';
 
 export {
   // Catch any errors thrown by the Layout component.
@@ -42,7 +41,7 @@ export default function RootLayout() {
     return null;
   }
 
-  return <Login />;
+  return <Onboarding />;
 }
 
 function RootLayoutNav() {
